Fix syntax error with previous fix.
The previous fix wasn't tested because I was pulling from github and
so didn't realize that the model parameter from the old method was
being used.  This is based on the code in
github.com/johncarney/validates_timeliness and is just as tested as
the last commit.  Will keep updating until it works properly.